# api/app.py
"""utilsim API. The contract is documented in docs/CONTRACT.md; OpenAPI at /openapi.json."""

from __future__ import annotations

import io
import os
from typing import Any, Literal

import numpy as np
import orjson
from fastapi import FastAPI, HTTPException, Query
from fastapi.middleware.cors import CORSMiddleware
from fastapi.middleware.gzip import GZipMiddleware
from fastapi.responses import Response

from api._agent import router as agent_router
from api._m2c import router as m2c_router
from api._ops import router as ops_router
from api._share import router as share_router
from api._store import store
from api._towns import health as health_body
from api._towns import ready_town as _town
from api._towns import router as towns_router
from utilsim.config import SCENARIOS
from utilsim.config.presets import deep_merge
from utilsim.io.geojson import LAYERS, layer
from utilsim.io.tables import table_rows, to_csv_text, to_parquet_bytes
from utilsim.process.fixtures import VARIANTS, fixture, variant
from utilsim.sim.flows import FlowModel
from utilsim.version import GENERATOR_VERSION

app = FastAPI(title="utilsim", version=GENERATOR_VERSION,
              description="Seeded utility-town engine: geography, networks, customers, simulation, meter-to-cash.")
app.add_middleware(GZipMiddleware, minimum_size=2048)
app.add_middleware(CORSMiddleware, allow_origins=os.environ.get("CORS_ORIGINS", "*").split(","),
                   allow_methods=["*"], allow_headers=["*"])
app.include_router(ops_router)  # operations
app.include_router(agent_router)
app.include_router(share_router)  # simulation codes: copy a simulation, paste it into another Utility Studio
app.include_router(m2c_router)  # meter-to-cash: reads, VEE, work queues
app.include_router(towns_router)  # generated towns (POST /api/towns), also a snapshot source for the two above


def J(data: Any, status: int = 200) -> Response:
    return Response(orjson.dumps(data, option=orjson.OPT_SERIALIZE_NUMPY), status_code=status,
                    media_type="application/json")


@app.get("/api/health")
def health():
    """Status, versions, towns (pack presets, then ready generated town ids, with details in ``generated``) and
    ``capabilities`` (``generate``: this engine can build new towns with ``POST /api/towns``)."""
    return J(health_body("local"))


@app.get("/api/schemas/{name}.json")
def get_published_schema(name: str):
    """Published JSON Schemas: utility-town-2.0, utility-state-1.0, utility-replay-1.0, meter-read-1.1,
    vee-input-fixture-1.1, config, openapi."""
    from utilsim.io.schemas import NAMES, load

    if name not in NAMES:
        raise HTTPException(404, f"unknown schema {name}; available: {sorted(NAMES)}")
    return J(load(name))


@app.get("/api/towns/{tid}/layers/{name}.geojson")
def get_layer(tid: str, name: str, crs: Literal["wgs84", "local"] = "wgs84"):
    town = _town(tid)
    if name not in LAYERS:
        raise HTTPException(404, f"unknown layer {name}; available: {LAYERS}")
    return J(layer(town, name, crs))


@app.get("/api/towns/{tid}/network/{utility}")
def get_network(tid: str, utility: Literal["electric", "gas", "water"]):
    town = _town(tid)
    net = town.networks[utility]
    return J({"commodity": utility, "sourceId": net.source_id, "stationId": net.station_id, "unit": net.unit,
              "nodes": {"id": [n.id for n in net.nodes], "kind": [n.kind for n in net.nodes],
                        "x": [round(float(n.xy[0]), 2) for n in net.nodes],
                        "z": [round(float(-n.xy[1]), 2) for n in net.nodes],
                        "parentEdge": [n.parent_edge for n in net.nodes]},
              "edges": {"id": [e.id for e in net.edges], "kind": [e.kind for e in net.edges],
                        "from": [net.nodes[e.a].id for e in net.edges], "to": [net.nodes[e.b].id for e in net.edges],
                        "tier": [e.tier for e in net.edges], "placement": [e.placement for e in net.edges],
                        "sizeMm": [int(e.size_mm) for e in net.edges],
                        "lengthM": [round(e.length, 2) for e in net.edges]},
              "meta": net.meta})


@app.get("/api/towns/{tid}/network/{utility}/trace/{node_id}")
def trace(tid: str, utility: Literal["electric", "gas", "water"], node_id: str, downstream_limit: int = 5000):
    town = _town(tid)
    net = town.networks[utility]
    try:
        v = net.index(node_id)
    except KeyError as exc:
        raise HTTPException(404, f"unknown node {node_id}") from exc
    up = []
    k = v
    while net.nodes[k].parent_edge >= 0:
        e = net.edges[net.nodes[k].parent_edge]
        up.append(e.id)
        k = e.a
    children: dict[int, list[int]] = {}
    for e in net.edges:
        children.setdefault(e.a, []).append(e.b)
    down, stack = [], [v]
    while stack and len(down) < downstream_limit:
        x = stack.pop()
        for c in children.get(x, []):
            down.append(net.nodes[c].id)
            stack.append(c)
    return J({"nodeId": node_id, "upstreamEdgeIds": up[::-1], "downstreamNodeIds": down,
              "downstreamTruncated": len(down) >= downstream_limit})


@app.get("/api/towns/{tid}/premises")
def list_premises(tid: str, mru: str | None = None, bbox: str | None = Query(None, description="minX,minZ,maxX,maxZ"),
                  limit: int = 1000, offset: int = 0):
    snap = store.snapshot(_town(tid).id)
    rows = snap["premises"]
    if mru:
        rows = [p for p in rows if p.get("mruId") == mru]
    if bbox:
        x0, z0, x1, z1 = (float(v) for v in bbox.split(","))
        rows = [p for p in rows if x0 <= p["x"] <= x1 and z0 <= p["z"] <= z1]
    keep = ("id", "address", "x", "z", "premiseType", "occupied", "solar", "services", "mruId", "accountId")
    return J({"total": len(rows), "items": [{k: p.get(k) for k in keep} for p in rows[offset:offset + limit]]})


@app.get("/api/towns/{tid}/premises/{pid}")
def premise_stack(tid: str, pid: str):
    snap = store.snapshot(_town(tid).id)
    prem = next((p for p in snap["premises"] if p["id"] == pid), None)
    if prem is None:
        raise HTTPException(404, f"unknown premise {pid}")
    sps = [s for s in snap["servicePoints"] if s["premiseId"] == pid]
    meter_ids = {s["meterId"] for s in sps}
    inst_ids = {s["installationId"] for s in sps}
    contracts = [c for c in snap["contracts"] if c["installationId"] in inst_ids]
    acc_ids = {c["accountId"] for c in contracts}
    accounts = [a for a in snap["accounts"] if a["id"] in acc_ids]
    bp_ids = {a["businessPartnerId"] for a in accounts}
    return J({"premise": prem, "building": next(b for b in snap["buildings"] if b["id"] == prem["buildingId"]),
              "parcel": next((p for p in snap["parcels"] if p["premiseId"] == pid), None),
              "servicePoints": sps, "installations": [i for i in snap["installations"] if i["id"] in inst_ids],
              "meters": [m for m in snap["meters"] if m["id"] in meter_ids],
              "registers": [r for r in snap["registers"] if r["meterId"] in meter_ids],
              "contracts": contracts, "accounts": accounts,
              "businessPartners": [b for b in snap["businessPartners"] if b["id"] in bp_ids],
              "tariffAssignments": [t for t in snap["tariffAssignments"] if t["contractId"] in
                                    {c["id"] for c in contracts}],
              "reads": [r for r in snap["sampleReads"] if r["premiseId"] == pid],
              "mru": next((m for m in snap["mrus"] if m["id"] == prem.get("mruId")), None)})


_flow_models: dict[str, FlowModel] = {}


@app.get("/api/towns/{tid}/flows")
def flows(tid: str, hour: float = Query(8.0, ge=0, lt=24), scenario: str = "normal", target: str | None = None):
    town = _town(tid)
    if scenario not in SCENARIOS:
        raise HTTPException(422, f"unknown scenario {scenario}")
    fm = _flow_models.get(tid) or _flow_models.setdefault(tid, FlowModel(town))
    res = fm.flows(hour, scenario, target or (town.prem.ids[0] if scenario == "leak" else None))
    return J({"hour": hour, "scenario": scenario, "source": res.source, "unit": res.unit,
              "edgeFlows": {u: np.round(v, 5) for u, v in res.edge_flows.items()},
              "edgeIds": {u: [e.id for e in town.networks[u].edges] for u in town.networks}})


_frame_builders: dict[str, Any] = {}


def _builder(tid: str):
    from utilsim.sim.state import FrameBuilder

    town = _town(tid)
    if tid not in _frame_builders:
        fm = _flow_models.get(tid) or _flow_models.setdefault(tid, FlowModel(town))
        _frame_builders[tid] = FrameBuilder(town, fm)
    return _frame_builders[tid]


@app.get("/api/towns/{tid}/state")
def state_frame(tid: str, hour: float = Query(8.0, ge=0, lt=24), date: str | None = None,
                scenario: str = "normal", target: str | None = None, premises: bool = True):
    """One complete ``utility-state/1.0`` frame (static solver output). Stateless and idempotent."""
    from utilsim.sim.state import local_time

    if scenario not in SCENARIOS:
        raise HTTPException(422, f"unknown scenario {scenario}")
    fb = _builder(tid)
    day = date or fb.town.cfg.scenario.date
    return J(fb.frame(local_time(fb.town, day, hour), scenario=scenario, target=target, include_premises=premises))


@app.get("/api/towns/{tid}/replay")
def replay(tid: str, date: str | None = None, scenario: str = "normal", target: str | None = None,
           start_hour: float = Query(0.0, ge=0, lt=24, alias="startHour"),
           hours: float = Query(24.0, gt=0, le=168), step_minutes: int = Query(60, ge=1, le=1440, alias="stepMinutes"),
           premises: bool = True):
    """``utility-replay/1.0``: frames with one simulationId and increasing sequence (≤ 2,000 frames)."""
    if scenario not in SCENARIOS:
        raise HTTPException(422, f"unknown scenario {scenario}")
    fb = _builder(tid)
    try:
        return J(fb.replay(date or fb.town.cfg.scenario.date, scenario=scenario, target=target,
                           start_hour=start_hour, hours=hours, step_minutes=step_minutes, include_premises=premises))
    except ValueError as exc:
        raise HTTPException(422, str(exc)) from exc


@app.get("/api/towns/{tid}/tables/{name}.{fmt}")
def get_table(tid: str, name: str, fmt: Literal["parquet", "csv", "json"]):
    snap = store.snapshot(_town(tid).id)
    try:
        rows = table_rows(snap, name)
    except KeyError as exc:
        raise HTTPException(404, str(exc)) from exc
    if fmt == "parquet":
        return Response(to_parquet_bytes(rows), media_type="application/vnd.apache.parquet")
    if fmt == "csv":
        return Response(to_csv_text(rows), media_type="text/csv")
    return J(rows)


@app.get("/api/towns/{tid}/render.png")
def get_render(tid: str, services: bool = False):
    from utilsim.io.render_png import render

    town = _town(tid)
    buf = io.BytesIO()
    path = render(town, f"{store_dir(tid)}/render{'-svc' if services else ''}.png", services=services)
    buf.write(path.read_bytes())
    return Response(buf.getvalue(), media_type="image/png")


def store_dir(tid: str) -> str:
    from api._store import CACHE_DIR

    d = CACHE_DIR / tid
    d.mkdir(parents=True, exist_ok=True)
    return str(d)


@app.get("/api/towns/{tid}/fixtures/vee.json")
def vee_fixture(tid: str, include_truth: bool = False):
    snap = store.snapshot(_town(tid).id)
    return J(fixture(snap["sampleReads"], include_truth))


@app.get("/api/towns/{tid}/fixtures/vee/{pid}/{commodity}.json")
def vee_fixture_one(tid: str, pid: str, commodity: Literal["electric", "gas", "water"],
                    variant_name: str = Query("actual", alias="variant"), include_truth: bool = False):
    if variant_name not in VARIANTS:
        raise HTTPException(422, f"variant must be one of {VARIANTS}")
    snap = store.snapshot(_town(tid).id)
    reads = [variant(r, variant_name) for r in snap["sampleReads"]
             if r["premiseId"] == pid and r["commodity"] == commodity]
    if not reads:
        raise HTTPException(404, f"no {commodity} reads for {pid}")
    return J(fixture(reads, include_truth))


_ = deep_merge