# schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Seeded utility-town engine: geography, networks, customers, simulation, meter-to-cash.",
    "title": "utilsim",
    "version": "0.10.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/schemas/{name}.json": {
      "get": {
        "description": "Published JSON Schemas: utility-town-2.0, utility-state-1.0, utility-replay-1.0, meter-read-1.1,\nvee-input-fixture-1.1, config, openapi.",
        "operationId": "get_published_schema_api_schemas__name__json_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Published Schema"
      }
    },
...
    "/api/towns/{tid}/fixtures/vee.json": {
      "get": {
        "operationId": "vee_fixture_api_towns__tid__fixtures_vee_json_get",
        "parameters": [
          {
            "in": "path",
            "name": "tid",
            "required": true,
            "schema": {
              "title": "Tid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_truth",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Truth",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Vee Fixture"
      }
    },
    "/api/towns/{tid}/fixtures/vee/{pid}/{commodity}.json": {
      "get": {
        "operationId": "vee_fixture_one_api_towns__tid__fixtures_vee__pid___commodity__json_get",
        "parameters": [
          {
            "in": "path",
            "name": "tid",
            "required": true,
            "schema": {
              "title": "Tid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "commodity",
            "required": true,
            "schema": {
              "enum": [
                "electric",
                "gas",
                "water"
              ],
              "title": "Commodity",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "variant",
            "required": false,
            "schema": {
              "default": "actual",
              "title": "Variant",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_truth",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Truth",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Vee Fixture One"
      }
    },
    "/api/towns/{tid}/flows": {
      "get": {
        "operationId": "flows_api_towns__tid__flows_get",
        "parameters": [
          {
            "in": "path",
            "name": "tid",
            "required": true,
            "schema": {
              "title": "Tid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hour",
            "required": false,
            "schema": {
              "default": 8.0,
              "exclusiveMaximum": 24,
              "minimum": 0,
              "title": "Hour",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "scenario",
            "required": false,
            "schema": {
              "default": "normal",
              "title": "Scenario",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Flows"
      }
    },
    "/api/towns/{tid}/layers/{name}.geojson": {
      "get": {
        "operationId": "get_layer_api_towns__tid__layers__name__geojson_get",
        "parameters": [
          {
            "in": "path",
            "name": "tid",
            "required": true,
            "schema": {
              "title": "Tid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "crs",
            "required": false,
            "schema": {
              "default": "wgs84",
              "enum": [
                "wgs84",
                "local"
              ],
              "title": "Crs",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Layer"
      }
    },
    "/api/towns/{tid}/network/{utility}": {
      "get": {
        "operationId": "get_network_api_towns__tid__network__utility__get",
        "parameters": [
          {
            "in": "path",
            "name": "tid",
            "required": true,
            "schema": {
              "title": "Tid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "utility",
            "required": true,
            "schema": {
              "enum": [
                "electric",
                "gas",
                "water"
              ],
              "title": "Utility",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Network"
      }
    },
    "/api/towns/{tid}/network/{utility}/trace/{node_id}": {
      "get": {
        "operationId": "trace_api_towns__tid__network__utility__trace__node_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "tid",
            "required": true,
            "schema": {
              "title": "Tid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "utility",
            "required": true,
            "schema": {
              "enum": [
                "electric",
                "gas",
                "water"
              ],
              "title": "Utility",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "downstream_limit",
            "required": false,
            "schema": {
              "default": 5000,
              "title": "Downstream Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trace"
      }
    },
    "/api/towns/{tid}/premises": {
      "get": {
        "operationId": "list_premises_api_towns__tid__premises_get",
        "parameters": [
          {
            "in": "path",
            "name": "tid",
            "required": true,
            "schema": {
              "title": "Tid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "mru",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mru"
            }
          },
          {
            "description": "minX,minZ,maxX,maxZ",
            "in": "query",
            "name": "bbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "minX,minZ,maxX,maxZ",
              "title": "Bbox"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 1000,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Premises"
      }
    },
    "/api/towns/{tid}/premises/{pid}": {
      "get": {
        "operationId": "premise_stack_api_towns__tid__premises__pid__get",
        "parameters": [
          {
            "in": "path",
            "name": "tid",
            "required": true,
            "schema": {
              "title": "Tid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Premise Stack"
      }
    },
    "/api/towns/{tid}/render.png": {
      "get": {
        "operationId": "get_render_api_towns__tid__render_png_get",
        "parameters": [
          {
            "in": "path",
            "name": "tid",
            "required": true,
            "schema": {
              "title": "Tid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "services",
            "required": false,
            "schema": {
              "default": false,
              "title": "Services",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Render"
      }
    },
    "/api/towns/{tid}/replay": {
      "get": {
        "description": "``utility-replay/1.0``: frames with one simulationId and increasing sequence (≤ 2,000 frames).",
        "operationId": "replay_api_towns__tid__replay_get",
        "parameters": [
          {
            "in": "path",
            "name": "tid",
            "required": true,
            "schema": {
              "title": "Tid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "in": "query",
            "name": "scenario",
            "required": false,
            "schema": {
              "default": "normal",
              "title": "Scenario",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target"
            }
          },
          {
            "in": "query",
            "name": "startHour",
            "required": false,
            "schema": {
              "default": 0.0,
              "exclusiveMaximum": 24,
              "minimum": 0,
              "title": "Starthour",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "hours",
            "required": false,
            "schema": {
              "default": 24.0,
              "exclusiveMinimum": 0,
              "maximum": 168,
              "title": "Hours",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "stepMinutes",
            "required": false,
            "schema": {
              "default": 60,
              "maximum": 1440,
              "minimum": 1,
              "title": "Stepminutes",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "premises",
            "required": false,
            "schema": {
              "default": true,
              "title": "Premises",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replay"
      }
    },
...
    "/api/towns/{tid}/state": {
      "get": {
        "description": "One complete ``utility-state/1.0`` frame (static solver output). Stateless and idempotent.",
        "operationId": "state_frame_api_towns__tid__state_get",
        "parameters": [
          {
            "in": "path",
            "name": "tid",
            "required": true,
            "schema": {
              "title": "Tid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hour",
            "required": false,
            "schema": {
              "default": 8.0,
              "exclusiveMaximum": 24,
              "minimum": 0,
              "title": "Hour",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "in": "query",
            "name": "scenario",
            "required": false,
            "schema": {
              "default": "normal",
              "title": "Scenario",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target"
            }
          },
          {
            "in": "query",
            "name": "premises",
            "required": false,
            "schema": {
              "default": true,
              "title": "Premises",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "State Frame"
      }
    },
...
  }
}